agent_runtime/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
azure-core==1.38.2
azure-functions==1.24.0
azure-servicebus==7.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
MarkupSafe==3.0.3
mysqlclient==2.2.8
openai==2.26.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
